Security refactoring
