+++ src/main/java/com/softserve/itacademy/vkhomenko/serviceapp2/exception/RegistrationException.java
@@ -1,0 +1,8 @@
+package com.softserve.itacademy.vkhomenko.serviceapp2.exception;
+
+public class RegistrationException extends RuntimeException {
+
+    public RegistrationException(String message) {
+        super(message);
+    }
+}
