+++ src/main/java/com/softserve/itacademy/vkhomenko/serviceapp2/repository/UserRepository.java
@@ -9,4 +9,5 @@ import java.util.UUID;
 public interface UserRepository extends JpaRepository<UserEntity, UUID> {
 
     Optional<UserEntity> findByEmail(String email);
+    boolean existsByEmail(String email);
 }
