+++ src/main/java/com/softserve/itacademy/vkhomenko/serviceapp2/security/JwtTokenFilter.java
@@ -1,5 +1,6 @@
 package com.softserve.itacademy.vkhomenko.serviceapp2.security;
 
+import com.softserve.itacademy.vkhomenko.serviceapp2.exception.JwtAuthException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
